# Conflicts:
#	cicd/deploy/prod.yaml
#	cicd/deploy/values.yaml

+++ integration/flow/src/test/kotlin/com/rarible/protocol/union/integration/flow/converter/FlowOrderConverterTest.kt
@@ -5,11 +5,12 @@ import com.rarible.protocol.union.dto.FlowAssetTypeFtDto
 import com.rarible.protocol.union.dto.FlowAssetTypeNftDto
 import com.rarible.protocol.union.dto.FlowOrderDataV1Dto
 import com.rarible.protocol.union.dto.PlatformDto
-import com.rarible.protocol.union.test.data.randomFlowV1OrderDto
+import com.rarible.protocol.union.integration.flow.data.randomFlowV1OrderDto
 import com.rarible.protocol.union.test.mock.CurrencyMock
 import kotlinx.coroutines.runBlocking
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
+import java.math.BigDecimal
 
 class FlowOrderConverterTest {
 
@@ -26,7 +27,7 @@ class FlowOrderConverterTest {
         assertThat(converted.startedAt).isNull()
         assertThat(converted.endedAt).isNull()
         // TODO makeStock is needed to be BigDecimal on the flow client side
-        assertThat(converted.makeStock).isEqualTo(dto.makeStock.toBigDecimal())
+        assertThat(converted.makeStock).isEqualTo(dto.makeStock)
         assertThat(converted.fill).isEqualTo(dto.fill)
         assertThat(converted.status.name).isEqualTo(dto.status!!.name)
         assertThat(converted.createdAt).isEqualTo(dto.createdAt)
